refactor: AttendanceHistoryLoader의 패키지 loader -> domain으로 변경
