+++ src/main/java/Application.java
@@ -1,6 +1,6 @@
 import controller.AttendanceController;
+import domain.AttendanceHistoryLoader;
 import domain.Crews;
-import loader.AttendanceHistoryLoader;
 import view.InputView;
 import view.OutputView;
